requirements.txt
annotated-types==0.6.0
anyio==3.7.1
boto3==1.26.10
botocore==1.29.165
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
fastapi==0.105.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
jmespath==1.0.1
loguru==0.7.2
pika==1.3.2
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.6.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
urllib3==1.26.18
uvicorn==0.24.0.post1
